+++ src/utils/SocketManager.h
@@ -18,8 +18,8 @@ public:
 
     int initSocket(const string &ip_address, int port, sockaddr_in& server_address, bool b);
     int accept();
-    int send(uint8_t *message_buffer, int message_buffer_size);
-    int receive(uint8_t *message_buffer, int message_buffer_size);
+    int send(uint8_t *message_buffer, size_t message_buffer_size);
+    int receive(uint8_t *message_buffer, size_t message_buffer_size);
 };
